backend/pyproject.toml
[project]
name = "simple-dspy-ui"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Christoph Holthaus", email = "christoph.holthaus+git@choltha.net"},
]
dependencies = [
    "dspy-ai>=2.4.9",
    "gradio>=4.36.1",
    "python-dotenv>=1.0.1",
    "debugpy>=1.8.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "-"}

[tool.pdm]
distribution = false
